refactor: Single analog clock one interval
